pyproject.toml
[tool.poetry]
name = "typeform-py"
version = "0.1.0"
description = "Simple CLI client to scroll through reddit posts"
authors = ["Noelle <73260931+No767@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
praw = "^7.7.0"
typer = {extras = ["all"], version = "^0.9.0"}
python-dotenv = "^1.0.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"


[tool.poetry.group.build.dependencies]
nuitka = "^1.5.7"

[tool.isort]
profile = 'black'

[tool.bandit]
skips = ["B311", "B101"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
